fix types for sdm_send() invocations

+++ sdmsh_commands.c
@@ -82,7 +82,7 @@ int sdmsh_cmd_config(struct shell_config *sc, char *argv[], int argc)
     if (argc == 5) {
         SDM_CHECK_STR_ARG_LONG("config: preamp gain", argv[4], preamp_gain, arg >= 0 && arg <= 13);
     }
-    sdm_send(ss, SDM_CMD_CONFIG, threshold, gain, srclvl, preamp_gain);
+    sdm_send(ss, SDM_CMD_CONFIG, (unsigned)threshold, (unsigned)gain, (unsigned)srclvl, (unsigned)preamp_gain);
     sdm_set_idle_state(ss);
 
     return 0;
@@ -100,7 +100,7 @@ int sdmsh_cmd_usbl_config(struct shell_config *sc, char *argv[], int argc)
     SDM_CHECK_STR_ARG_LONG("usbl_config: gain", argv[3], gain, arg >= 0 && arg <= 13);
     SDM_CHECK_STR_ARG_LONG("usbl_config: sample_rate", argv[4], sample_rate, arg >= 0 && arg <= 6);
 
-    sdm_send(ss, SDM_CMD_USBL_CONFIG, delay, samples, gain, sample_rate);
+    sdm_send(ss, SDM_CMD_USBL_CONFIG, (unsigned)delay, (unsigned)samples, (unsigned)gain, (unsigned)sample_rate);
     sdm_set_idle_state(ss);
 
     return 0;
@@ -162,7 +162,7 @@ int sdmsh_cmd_ref(struct shell_config *sc, char *argv[], int argc)
         }
 
         /* DUMP2LOG(ERR_LOG, (char *)data, len*2); */
-        rc = sdm_send(ss, SDM_CMD_REF, data, len);
+        rc = sdm_send(ss, SDM_CMD_REF, data, (unsigned)len);
     }
 
     stream_close(stream);
@@ -235,11 +235,11 @@ int sdmsh_cmd_tx(struct shell_config *sc, char *argv[], int argc)
         logger(DATA_LOG, "tx data %d / %d / %d samples  \r", nsamples, len, passed);
 
         if (cnt == len && nsamples >= 1024) {
-            rc = sdm_send(ss, cmd, nsamples, data, len);
+            rc = sdm_send(ss, cmd, (unsigned)nsamples, data, (unsigned)len);
             passed += len;
         } else if (cnt > 0) {
             memset(&data[cnt], 0, (len - cnt) * 2);
-            rc = sdm_send(ss, cmd, nsamples, data, 1024 * ((cnt + 1023) / 1024));
+            rc = sdm_send(ss, cmd, (unsigned)nsamples, data, (unsigned)(1024 * ((cnt + 1023) / 1024)));
             passed += 1024 * ((cnt + 1023) / 1024);
         } else {
             rc = -1;
@@ -301,7 +301,7 @@ int sdmsh_cmd_rx_helper(struct shell_config *sc, char *argv[], int argc, int cod
         return -1;
     }
 
-    sdm_send(ss, code, nsamples);
+    sdm_send(ss, code, (unsigned long)nsamples);
     /* rl_message("Waiting for receiving %ld samples to file %s\n", nsamples, ss->filename); */
     
     return 0;
@@ -340,7 +340,7 @@ int sdmsh_cmd_usbl_rx(struct shell_config *sc, char *argv[], int argc)
             logger(ERR_LOG, "usbl_rx: error %s\n", stream_strerror(stream));
         return -1;
     }
-    sdm_send(ss, SDM_CMD_USBL_RX, channel, samples);
+    sdm_send(ss, SDM_CMD_USBL_RX, (unsigned)channel, (unsigned)samples);
     
     return 0;
 }
